% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagear_data.R
\docType{data}
\name{example_references_metagear}
\alias{example_references_metagear}
\title{A collection of bibliographic references}
\format{A data frame with 12 rows and 9 variables.}
\usage{
data(example_references_metagear)
}
\description{
An example dataset containing 11 journal references.  The variables are
described below.
}
\details{
\itemize{
  \item AUTHORS. Authors of the journal article
  \item YEAR. Publication year
  \item TITLE. Article title
  \item JOURNAL. Journal name
  \item VOLUME. Journal volume number
  \item LPAGES. Lower page number
  \item UPAGES. Upper page number
  \item DOI. Digital object identifier (DOI) of journal article
  \item ABSTRACT. Full text of the journal article abstract
}
}
\keyword{datasets}

