% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figure_displayDetectedPoints.R
\name{figure_displayDetectedPoints}
\alias{figure_displayDetectedPoints}
\title{Displays detected points on figure.}
\usage{
figure_displayDetectedPoints(aDetectedPlot, background = NULL,
  color = "red", size = 2, ignore = FALSE)
}
\arguments{
\item{aDetectedPlot}{A binary figure image with detected points
(an \code{EBImage} object).  See: \code{\link{figure_detectAllPoints}}}

\item{background}{An EBImage figure of same size to be used as
background (e.g., the original (RGB/color) figure image).}

\item{color}{The color to paint the detected points.}

\item{size}{The radius of the painted points.}

\item{ignore}{When \code{TRUE} does not display painted image, only
   returns painted image EBImage object.}
}
\value{
A RGB \code{EBImage} painted with detected points.
}
\description{
Generates a raster image of a figure with the detected points painted on a
   background/reference figure.
}
\seealso{
\link{figure_displayDetections}
}

