% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_displayDetections.R
\name{figure_displayDetections}
\alias{figure_displayDetections}
\title{Displays the detected figure objects.}
\usage{
figure_displayDetections(aDetectedPlot, background = NULL, color = "red",
  ignore = FALSE)
}
\arguments{
\item{aDetectedPlot}{A binary figure image with detected objects
(an \code{EBImage} object).}

\item{background}{An \code{EBImage} figure of same size to be used as
background (e.g., the original [RGB/color] figure image).}

\item{color}{The color to paint the detected objects.}

\item{ignore}{When \code{TRUE} does not display painted image, only 
returns painted image EBImage object.}
}
\value{
A RGB \code{EBImage} painted with detected figure objects.
}
\description{
Generates a raster image of a figure with the detected objects painted on a 
   background/reference figure.
}
