% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_read.R
\name{figure_read}
\alias{figure_read}
\title{Reads/loads a figure image from file.}
\usage{
figure_read(file = file.choose(), display = FALSE)
}
\arguments{
\item{file}{The file name and location of a plot figure.  Prompts
for file name if none is explicitly called.  Preferably in .jpg format.}

\item{display}{When \code{"TRUE"}, displays the figure as a raster image.}
}
\value{
An \code{EBImage} object figure.
}
\description{
Reads a .jpg, .jpeg, .png, or .tiff image file containing a plotted figure.
}
\seealso{
\link{figure_write}
}
