% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/figure_removeOutlyingPoints.R
\name{figure_removeOutlyingPoints}
\alias{figure_removeOutlyingPoints}
\title{Remove outlier points from a figure.}
\usage{
figure_removeOutlyingPoints(aDetectedPlot, xAxis = NULL, yAxis = NULL)
}
\arguments{
\item{aDetectedPlot}{A binary figure image with detected points (an
\code{EBImage} object).
See: \code{\link{figure_detectAllPoints}}}

\item{xAxis}{A binary figure image with detected X-axis (an \code{EBImage} object).
See: \code{\link{figure_detectAxis}}}

\item{yAxis}{A binary figure image with detected Y-axis (an \code{EBImage} object).
   See: \code{\link{figure_detectAxis}}}
}
\value{
An \code{EBImage} object with detected points within the specified X-
   and Y-axis ranges.
}
\description{
Removes all detected points outside of axis range.  Requires three detected
images: one based on \code{\link{figure_detectAllPoints}}, and two others based on
detected X- and Y-axes (i.e. \code{\link{figure_detectAxis}})
}

