% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_write.R
\name{figure_write}
\alias{figure_write}
\title{Saves/writes a figure image.}
\usage{
figure_write(aFigure, file = NULL)
}
\arguments{
\item{aFigure}{The \code{EBImage} figure.}

\item{file}{Name and location of file to save.  Supports .jpg, .png, and
.tiff image formats.}
}
\value{
Vector of file names.
}
\description{
Writes a figure image to file and returns the file name.
}
\seealso{
\code{\link{figure_read}}
}

