% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impute_missingness.R
\name{impute_missingness}
\alias{impute_missingness}
\title{Provides a summary of missingness in a dataset.}
\usage{
impute_missingness(aDataFrame)
}
\arguments{
\item{aDataFrame}{A data.frame containing columns that will be assessed for
   missignness.}
}
\value{
A data frame that summarizes percent missingness for each column of
   a dataset.
}
\description{
Generates a summary of the percentage of missing data in a dataset.  Provides
insight on the appropriateness of imputation methods.  For example, if 30\%
of data is missing, then perhaps this is too much to impute.
}
\examples{
data(example_references_metagear)
impute_missingness(example_references_metagear)
}

