% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/isPDF.R
\name{isPDF}
\alias{isPDF}
\title{Evaluates whether a file is a PDF document.}
\usage{
isPDF(aFileName)
}
\arguments{
\item{aFileName}{A string that identifies a file name (and direcotry path) of
the PDF candidate.}
}
\value{
A logical value indicating whether the file is a PDF document.
}
\description{
Checks if provided file is in Portable Document Format (PDF).
}

