% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagear_data.R
\docType{data}
\name{Kam_et_al_2003_Fig2}
\alias{Kam_et_al_2003_Fig2}
\title{An example image of a scatterplot figure}
\format{A \code{EBImage} object containing a scatterplot raster image}
\usage{
data(Kam_et_al_2003_Fig2)
}
\description{
An EBImage object depicting a scatterplot image from Figure 2 of Kam, M., 
   Cohen-Gross, S., Khokhlova, I.S., Degen, A.A. and Geffen, E. 2003. 
   Average daily metabolic rate, reproduction and energy allocation during
   lactation in the Sundevall Jird Meriones crassus. Functional 
   Ecology 17:496-503.
}
\keyword{datasets}

