% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metagear_data.R
\docType{data}
\name{Kortum_and_Acymyan_2013_Fig4}
\alias{Kortum_and_Acymyan_2013_Fig4}
\title{An example image of a bar plot figure}
\format{A \code{EBImage} object containing a bar plot raster image}
\usage{
data(Kortum_and_Acymyan_2013_Fig4)
}
\description{
An EBImage object depicting a bar plot image from Figure 4 of Kortum, P., and
   Acymyan, C.Z. 2013. How low can you go? Is the System Usability Scale
   range restricted? Journal of Usability Studies 9:14-24.
}
\keyword{datasets}

