% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDF_extractImages.R
\name{PDF_extractImages}
\alias{PDF_extractImages}
\title{Attempts to extract all images from a PDF}
\usage{
PDF_extractImages(file = file.choose())
}
\arguments{
\item{file}{The file name and location of a PDF file.  Prompts
for file name if none is explicitly called.}
}
\value{
A vector of file names saved as images.
}
\description{
Tries to extract images within a PDF file.  Currently does not support 
decoding of images in CCITT compression formats. However, will still save 
these images to file; as a record of the number of images detected in the PDF.
}

