% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_phyloMeta1.3.R
\name{replicate_phyloMeta1.3}
\alias{replicate_phyloMeta1.3}
\title{Replicate phylogeneic meta-analysis results and summaries from phyloMeta 1.3.}
\usage{
replicate_phyloMeta1.3(model, weights, data, phylogenyFile)
}
\arguments{
\item{model}{A two-sided linear formula object describing the model, with the
response (effect sizes) on the left of a ~ operator and the moderator
variables, separated by +, :, * operators, on the right.  NOTE:
phyloMeta was limited to analyses with a single moderator variable.
This function currently supports only numerical categorical moderators.}

\item{weights}{A vector of effect size variances that will be used as
weights for the meta-analysis.}

\item{data}{A data frame containing the variables named in model
and weights and species names (names must be exact as specified in
phylogeny).}

\item{phylogenyFile}{A text file containing a NEWICK phylogeny. The number of
species must be same as (k) number of effect sizes in data.}
}
\description{
Replicate phylogenetic meta-analysis results and summaries from Lajeunesse
(2011) software 'phyloMeta' 1.3.  Currently does not fully replicate all
functionality.
}
\references{
Lajeunesse, M.J. (2011) phyloMeta: a program for phylogenetic
   comparative analyses with meta-analysis. Bioinformatics 27, 2603-2604.
}
